Rearrange serializer package structure
