--- extensions/serialization/src/commonMain/kotlin/io/islandtime/extensions/serialization/Date.kt
@@ -1,23 +1,0 @@
-package io.islandtime.extensions.serialization
-
-import io.islandtime.Date
-import io.islandtime.toDate
-import kotlinx.serialization.KSerializer
-import kotlinx.serialization.descriptors.PrimitiveKind
-import kotlinx.serialization.descriptors.PrimitiveSerialDescriptor
-import kotlinx.serialization.descriptors.SerialDescriptor
-import kotlinx.serialization.encoding.Decoder
-import kotlinx.serialization.encoding.Encoder
-
-object DateSerializer : KSerializer<Date> {
-    override val descriptor: SerialDescriptor =
-        PrimitiveSerialDescriptor("io.islandtime.DateSerializer", PrimitiveKind.STRING)
-
-    override fun serialize(encoder: Encoder, value: Date) {
-        encoder.encodeString(value.toString())
-    }
-
-    override fun deserialize(decoder: Decoder): Date {
-        return decoder.decodeString().toDate()
-    }
-}
