--- extensions/serialization/src/commonMain/kotlin/io/islandtime/extensions/serialization/DateTime.kt
@@ -1,23 +1,0 @@
-package io.islandtime.extensions.serialization
-
-import io.islandtime.DateTime
-import io.islandtime.toDateTime
-import kotlinx.serialization.KSerializer
-import kotlinx.serialization.descriptors.PrimitiveKind
-import kotlinx.serialization.descriptors.PrimitiveSerialDescriptor
-import kotlinx.serialization.descriptors.SerialDescriptor
-import kotlinx.serialization.encoding.Decoder
-import kotlinx.serialization.encoding.Encoder
-
-object DateTimeSerializer : KSerializer<DateTime> {
-    override val descriptor: SerialDescriptor =
-        PrimitiveSerialDescriptor("io.islandtime.DateTimeSerializer", PrimitiveKind.STRING)
-
-    override fun serialize(encoder: Encoder, value: DateTime) {
-        encoder.encodeString(value.toString())
-    }
-
-    override fun deserialize(decoder: Decoder): DateTime {
-        return decoder.decodeString().toDateTime()
-    }
-}
