--- extensions/serialization/src/commonMain/kotlin/io/islandtime/extensions/serialization/Instant.kt
@@ -1,23 +1,0 @@
-package io.islandtime.extensions.serialization
-
-import io.islandtime.Instant
-import io.islandtime.toInstant
-import kotlinx.serialization.KSerializer
-import kotlinx.serialization.descriptors.PrimitiveKind
-import kotlinx.serialization.descriptors.PrimitiveSerialDescriptor
-import kotlinx.serialization.descriptors.SerialDescriptor
-import kotlinx.serialization.encoding.Decoder
-import kotlinx.serialization.encoding.Encoder
-
-object InstantSerializer : KSerializer<Instant> {
-    override val descriptor: SerialDescriptor =
-        PrimitiveSerialDescriptor("io.islandtime.InstantSerializer", PrimitiveKind.STRING)
-
-    override fun serialize(encoder: Encoder, value: Instant) {
-        encoder.encodeString(value.toString())
-    }
-
-    override fun deserialize(decoder: Decoder): Instant {
-        return decoder.decodeString().toInstant()
-    }
-}
