--- extensions/serialization/src/commonMain/kotlin/io/islandtime/extensions/serialization/OffsetTime.kt
@@ -1,23 +1,0 @@
-package io.islandtime.extensions.serialization
-
-import io.islandtime.OffsetTime
-import io.islandtime.toOffsetTime
-import kotlinx.serialization.KSerializer
-import kotlinx.serialization.descriptors.PrimitiveKind
-import kotlinx.serialization.descriptors.PrimitiveSerialDescriptor
-import kotlinx.serialization.descriptors.SerialDescriptor
-import kotlinx.serialization.encoding.Decoder
-import kotlinx.serialization.encoding.Encoder
-
-object OffsetTimeSerializer : KSerializer<OffsetTime> {
-    override val descriptor: SerialDescriptor =
-        PrimitiveSerialDescriptor("io.islandtime.OffsetTimeSerializer", PrimitiveKind.STRING)
-
-    override fun serialize(encoder: Encoder, value: OffsetTime) {
-        encoder.encodeString(value.toString())
-    }
-
-    override fun deserialize(decoder: Decoder): OffsetTime {
-        return decoder.decodeString().toOffsetTime()
-    }
-}
