--- extensions/serialization/src/commonMain/kotlin/io/islandtime/extensions/serialization/measures/Duration.kt
@@ -1,23 +1,0 @@
-package io.islandtime.extensions.serialization.measures
-
-import io.islandtime.measures.Duration
-import io.islandtime.measures.toDuration
-import kotlinx.serialization.KSerializer
-import kotlinx.serialization.descriptors.PrimitiveKind
-import kotlinx.serialization.descriptors.PrimitiveSerialDescriptor
-import kotlinx.serialization.descriptors.SerialDescriptor
-import kotlinx.serialization.encoding.Decoder
-import kotlinx.serialization.encoding.Encoder
-
-object DurationSerializer : KSerializer<Duration> {
-    override val descriptor: SerialDescriptor =
-        PrimitiveSerialDescriptor("io.islandtime.measures.DurationSerializer", PrimitiveKind.STRING)
-
-    override fun serialize(encoder: Encoder, value: Duration) {
-        encoder.encodeString(value.toString())
-    }
-
-    override fun deserialize(decoder: Decoder): Duration {
-        return decoder.decodeString().toDuration()
-    }
-}
